# server/server/common/api.yml
openapi: 3.0.0
info:
  title: Blog API
  description: API for managing blogs and posts
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
...
  /api/blogs/{blogId}:
    get:
      summary: Get a specific blog by ID
      parameters:
        - in: path
          name: blogId
          required: true
          schema:
            type: string
          description: ID of the blog to get
      responses:
        '200':
          description: A blog object
        '404':
          description: Blog not found
    put:
      summary: Update a blog
      parameters:
        - in: path
          name: blogId
          required: true
          schema:
            type: string
          description: ID of the blog to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlogInput'
      responses:
        '200':
          description: Blog updated successfully
        '400':
          description: Invalid request body
        '404':
          description: Blog not found
    delete:
      summary: Delete a blog
      parameters:
        - in: path
          name: blogId
          required: true
          schema:
            type: string
          description: ID of the blog to delete
      responses:
        '204':
          description: Blog deleted successfully
        '404':
          description: Blog not found
...
components:
  schemas:
    BlogInput:
      type: object
      properties:
        title:
          type: string
          description: Title of the blog
...
security:
  - JWT: []